fixed all -Wsign-compare warning messages
Import internal commit 458d26f780774de442528dad51177a804311ba43

+++ groups/bsl/bslalg/bslalg_containerbase.t.cpp
@@ -584,7 +584,7 @@ int main(int argc, char *argv[])
             bslma::TestAllocatorMonitor tam(&ta);
 
             Obj::pointer ptr = mX.allocate(i);
-            ASSERTV(i, i * sizeof(TestType) == ta.numBytesInUse());
+            ASSERTV(i, i * sizeof(TestType) == (size_t) ta.numBytesInUse());
 
             int nc = TestType::numCopyConstruct();
             mX.construct(ptr, TestType());
@@ -603,7 +603,7 @@ int main(int argc, char *argv[])
 
             int *intPtr = 0;
             intPtr = mX.allocateN(intPtr, i);
-            ASSERTV(i, i * sizeof(int) == ta.numBytesInUse());
+            ASSERTV(i, i * sizeof(int) == (size_t) ta.numBytesInUse());
 
             mX.deallocateN(intPtr, i);
             ASSERTV(i, 0 == ta.numBytesInUse());
